Add auto-reconnection for gws mode
Ref #110
Ref #144
Ref #268
