+++ .github/workflows/test.yml
@@ -34,6 +34,30 @@ jobs:
             go-version: '1.24.1'
             connection-type: 'http'
             surrealdb-url: 'http://localhost:8000'
+          # v2.3.7 with Go 1.24.1
+          - surrealdb-version: 'v2.3.7'
+            go-version: '1.24.1'
+            connection-type: 'ws'
+            surrealdb-url: 'ws://localhost:8000/rpc'
+          # v2.3.7 with Go 1.24.1 with reconnection
+          - surrealdb-version: 'v2.3.7'
+            go-version: '1.24.1'
+            connection-type: 'ws'
+            surrealdb-url: 'ws://localhost:8000/rpc'
+            surrealdb-reconnection-check-interval: '1s'
+          # v2.3.7 with Go 1.24.1 with gws
+          - surrealdb-version: 'v2.3.7'
+            go-version: '1.24.1'
+            connection-type: 'http'
+            surrealdb-url: 'http://localhost:8000'
+            surrealdb-connection-impl: 'gws'
+          # v2.3.7 with Go 1.24.1 with gws and reconnection
+          - surrealdb-version: 'v2.3.7'
+            go-version: '1.24.1'
+            connection-type: 'http'
+            surrealdb-url: 'http://localhost:8000'
+            surrealdb-connection-impl: 'gws'
+            surrealdb-reconnection-check-interval: '1s'
           # v2.3.7 with Go 1.23.11
           - surrealdb-version: 'v2.3.7'
             go-version: '1.23.11'
@@ -79,3 +103,5 @@ jobs:
         run: go test -v -cover ./...
         env:
           SURREALDB_URL: ${{ matrix.surrealdb-url }}
+          SURREALDB_CONNECTION_IMPL: ${{ matrix.surrealdb-connection-impl }}
+          SURREALDB_RECONNECTION_CHECK_INTERVAL: ${{ matrix.surrealdb-reconnection-check-interval }}
