+++ pkg/connection/gorillaws/ws_test.go
@@ -1,4 +1,4 @@
-package connection
+package gorillaws
 
 import (
 	"testing"
